feat: add service-registry-api package and replace messages
