+++ Cargo.toml
@@ -58,6 +58,7 @@ schemars = "0.8.10"
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 serde_json = "1.0.89"
 service-registry = { version = "^1.0.0", path = "contracts/service-registry" }
+service-registry-api = { version = "^1.0.0", path = "packages/service-registry-api" }
 sha3 = { version = "0.10.8", default-features = false, features = [] }
 signature-verifier-api = { version = "^1.0.0", path = "packages/signature-verifier-api" }
 stellar = { version = "^1.0.0", path = "external-gateways/stellar" }
