+++ contracts/service-registry/src/bin/schema.rs
@@ -1,5 +1,6 @@
 use cosmwasm_schema::write_api;
-use service_registry::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use service_registry::msg::InstantiateMsg;
+use service_registry_api::msg::{ExecuteMsg, QueryMsg};
 
 fn main() {
     write_api! {
