+++ packages/service-registry-api/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "service-registry-api"
+version = "1.0.0"
+rust-version = { workspace = true }
+edition = { workspace = true }
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+axelar-wasm-std = { workspace = true, features = ["derive"] }
+cosmwasm-schema = { workspace = true }
+cosmwasm-std = { workspace = true }
+error-stack = { workspace = true }
+msgs-derive = { workspace = true }
+report = { workspace = true }
+router-api = { workspace = true }
+schemars = "0.8.10"
+serde = { version = "1.0.145", default-features = false, features = ["derive"] }
+
+[dev-dependencies]
+goldie = { workspace = true }
+
+[lints]
+workspace = true
