+++ packages/service-registry-api/release.toml
@@ -1,0 +1,1 @@
+release = false
