+++ packages/service-registry-api/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod msg;
